geotiff tests: consolidate GPU writer cluster (Sub-PR B)
Closes part of #2438 (Sub-PR B). Cluster 14 of long-tail epic #2424.

Fold 8 GPU writer test files into a new
xrspatial/geotiff/tests/gpu/test_writer.py:

- test_gpu_writer_attrs_1563.py (write-path attrs parity)
- test_gpu_writer_band_first_1580.py ((band, y, x) layout remap)
- test_gpu_writer_compression_modes_2026_05_11.py (zstd, deflate, jpeg, none)
- test_gpu_writer_cpu_fallback_codecs_2026_05_12.py (lzw, packbits, lz4, lerc, jpeg2000)
- test_gpu_writer_nan_sentinel_1599.py (NaN-to-sentinel rewrite)
- test_gpu_writer_overview_inplace_1948.py (cupy.putmask in COG loop)
- test_gpu_writer_overview_mode_and_compression_level_1740.py
- test_to_geotiff_gpu_fallback_1674.py (gpu=True exception classification)

Tests-only. The shared requires_gpu marker from _helpers/markers.py
replaces the per-file _gpu_available / _gpu_only helpers. 87 tests
collected, matching the baseline sum across the eight source files.

The release-gate row for writer.gpu and the geotiff.rst experimental
codec note now cite gpu/test_writer.py.
